requirements.txt
h5py==2.9.0
imageio==2.5.0
ipdb==0.12
ipython==7.5.0
matplotlib==3.1.0
numpy==1.16.4
pandas==0.24.2
Pillow==6.0.0
pyaml==19.4.1
PyYAML==5.1.1
scikit-learn==0.21.2
scipy==1.3.0
seaborn==0.9.0
sklearn==0.0
tensorboardX==1.7
tensorflow-gpu==1.14.0
torch==1.1.0
torchvision==0.3.0
tqdm==4.19.9
